chore: add custody menu
